PCB-Deffects-Detector-main/AOI_System/requirements.txt.txt
# AI & Computer Vision
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0

# Data Handling & Graphs
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# Hardware Communication
pyserial>=3.5

# GUI (Tkinter e built-in, dar uneori cere pillow-tk)
# Nota: Tkinter vine de obicei preinstalat cu Python

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
scipy>=1.10.0